+++ flake.nix
@@ -27,6 +27,7 @@
           pkgs = import nixpkgs {
             inherit system;
             overlays = [
+              (import ./nix/build_overlay.nix)
               gomod2nix.overlays.default
               self.overlay
             ];
@@ -46,7 +47,7 @@
           devShells = {
             cronosd = pkgs.mkShell {
               buildInputs = with pkgs; [
-                go_1_19
+                go_1_20
                 rocksdb
                 gomod2nix
               ];
@@ -58,13 +59,6 @@
       )
     ) // {
       overlay = final: super: {
-        go_1_19 = super.go_1_19.overrideAttrs (_: rec {
-          version = "1.19.6";
-          src = final.fetchurl {
-            url = "https://go.dev/dl/go${version}.src.tar.gz";
-            hash = "sha256-1/ABP4Lm1/hizGy1yM20ju9fLiObNbqpfi8adGYEN2c=";
-          };
-        });
         bundle-exe = final.pkgsBuildBuild.callPackage nix-bundle-exe { };
         # make-tarball don't follow symbolic links to avoid duplicate file, the bundle should have no external references.
         # reset the ownership and permissions to make the extract result more normal.
@@ -77,9 +71,6 @@
             | gzip -9 > $out
         '';
         bundle-win-exe = drv: final.callPackage ./nix/bundle-win-exe.nix { cronosd = drv; };
-        # only enable jemalloc for non-windows platforms
-        # see: https://github.com/NixOS/nixpkgs/issues/216479
-        rocksdb = final.callPackage ./nix/rocksdb.nix { enableJemalloc = !final.stdenv.hostPlatform.isWindows; };
       } // (with final;
         let
           matrix = lib.cartesianProductOfSets {