+++ integration_tests/test_filters.py
@@ -49,7 +49,7 @@ def test_event_log_filter(cronos):
         fromBlock=current_height
     )
 
-    tx = mycontract.functions.setGreeting("world").buildTransaction()
+    tx = mycontract.functions.setGreeting("world").build_transaction()
     tx_receipt = send_transaction(w3, tx)
     log = mycontract.events.ChangeGreeting().processReceipt(tx_receipt)[0]
     assert log["event"] == "ChangeGreeting"
@@ -59,3 +59,9 @@ def test_event_log_filter(cronos):
     print(f"get event: {new_entries}")
     assert new_entries[0] == log
     assert "world" == mycontract.caller.greet()
+    # without new txs since last call
+    assert event_filter.get_new_entries() == []
+    assert event_filter.get_all_entries() == new_entries
+    # Uninstall
+    assert w3.eth.uninstall_filter(event_filter.filter_id)
+    assert not w3.eth.uninstall_filter(event_filter.filter_id)